requirements.txt
# Core Framework
streamlit==1.32.0
streamlit-lottie==0.1.3

# Data Processing
pandas==2.2.1
numpy==1.24.3

# NLP & Machine Learning
transformers==4.40.0
torch==2.2.1
sentencepiece==0.2.0
spacy==3.7.4
nltk==3.8.1

# Visualization
plotly==5.19.0
matplotlib==3.8.3
wordcloud==1.9.2

# PDF Generation
fpdf2==2.7.4  # Modern, actively maintained version of fpdf

# Database
psycopg2-binary==2.9.9

# Cryptography for Authentication
bcrypt==4.1.3
pyotp==2.9.0

# Optional (for Excel files)
openpyxl==3.1.2